requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0
flasgger>=0.9.7
